写各种类型absPorts的筛选重载方法;

+++ SMG_NothingIsAll/AIFoundation/AIThinkingControl/ThinkingUtils.m
@@ -69,7 +69,7 @@
         return DefaultDataSource;
     }else{
         NSInteger value = [self getAnalogyTypeValue:type];
-        return STRFORMAT(@"%d",value);
+        return STRFORMAT(@"%ld",(long)value);
     }
     return DefaultDataSource;
 }
